+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Changelog
 
 
 - Add ``core.make_identity`` factory function for easily creating basic identity functions from a list of model column objects that can be used with ``save()``.
+- Import ``core.save``, ``core.destroy``, ``core.transaction``, and ``core.make_identity`` into make package namespace.
 
 
 v0.6.3 (2016-10-17)